Fixed the Google Button Redirect
