+++ frontend/Active-learning-project/.gitignore
@@ -6,6 +6,8 @@
 /out-tsc
 /bazel-out
 
+../ALP
+
 # Node
 /node_modules
 npm-debug.log
