+++ frontend/Active-learning-project/src/app/alp-app/app.component.spec.ts
