+++ frontend/Active-learning-project/src/app/master-layout/layout/master-layout.component.html
