+++ frontend/Active-learning-project/src/app/user-authentication/account/account.component.css
@@ -1,7 +1,7 @@
 .auth-container {
   position: relative;
   height: 100vh;
-  padding: 2rem 3rem;
+  padding: 1rem 3rem;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -13,7 +13,7 @@
   width: 100%;
   height: 100%;
   margin: 0;
-  border-radius: 0.6rem;
+  border-radius: 0.3rem;
   overflow: hidden;
   background-color: var(--bg-auth-container-color);
   box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
@@ -26,7 +26,7 @@
 .auth-background {
   position: relative;
   overflow: hidden;
-  background-image: url(../../assets/images/learn.jpg);
+  background-image: url(../../../assets/images/learn.jpg);
   background-position: top;
   background-repeat: no-repeat;
   background-size: cover;
