+++ frontend/Active-learning-project/src/app/user-authentication/account/account.component.html
@@ -5,7 +5,6 @@
         <router-outlet></router-outlet>
       </div>
       <div class="col-lg auth-background">
-        <!-- <video src="../../assets/videos/authbg.mp4" autoplay loop></video> -->
         <p>LEARNING SHOULD EASILY BE EFFECTIVE</p>
       </div>
     </div>
