+++ frontend/Active-learning-project/src/app/user-authentication/google-button/google-button.component.css
@@ -1,0 +1,3 @@
+.asl-google-signin-button{
+    border-radius: 10px !important;
+}
