+++ frontend/Active-learning-project/src/app/user-authentication/google-button/google-button.component.html
@@ -1,30 +1,6 @@
-<div *ngIf="btnType == 'signin'">
-  <div
-    id="g_id_onload"
-    data-client_id="642687867582-551qek22bgc3r2rumue30efe3b1arpeb.apps.googleusercontent.com"
-    data-context="signin"
-    data-ux_mode="popup"
-    data-login_uri="http://localhost:4200"
-    data-auto_prompt="false"
-  ></div>
-</div>
-<div *ngIf="btnType == 'signup'">
-  <div
-    id="g_id_onload"
-    data-client_id="642687867582-551qek22bgc3r2rumue30efe3b1arpeb.apps.googleusercontent.com"
-    data-context="signup"
-    data-ux_mode="popup"
-    data-login_uri="http://localhost:4200" 
-    data-auto_prompt="false"
-  ></div>
-</div>
-
-<div
-  class="g_id_signin"
-  data-type="standard"
-  data-shape="pill"
-  data-theme="outline"
-  data-text="signin_with"
-  data-size="large"
-  data-logo_alignment="left"
-></div>
+<asl-google-signin-button
+  class="googlebtn"
+  type="standard"
+  size="large"
+  class="google-auth-btn"
+></asl-google-signin-button>
