+++ frontend/Active-learning-project/src/app/user-authentication/signup/signup.component.html
@@ -36,7 +36,6 @@
           <div *ngIf="controls['firstName'].errors['required']">
             First name is required
           </div>
-        
         </div>
       </div>
       <div class="form-group">
@@ -103,10 +102,7 @@
         </div>
       </div>
       <div class="form-group auth-login">
-        <button
-          [disabled]="loading"
-          class="btn btn-dark form-control"
-        >
+        <button [disabled]="loading" class="btn btn-dark form-control">
           <span
             *ngIf="loading"
             class="spinner-border spinner-border-sm mr-1"
@@ -129,10 +125,7 @@
       </div>
     </div>
     <div class="auth-signup-link text-center">
-      <i
-        >Already have an account?
-        <a href="/auth/user/signin">Sign In</a></i
-      >
+      <i>Already have an account? <a [routerLink]="['/signin']">Sign In</a></i>
     </div>
   </div>
 </div>
